Changed doubles to decimals for more precision

+++ SFSimulator.Core/Characters/Character.cs
@@ -5,7 +5,7 @@
         public int Level { get; set; } = 400;
         public long Experience { get; set; } = 0;
         public int BaseStat { get; set; } = 0;
-        public float Gold { get; set; } = 0;
+        public decimal Gold { get; set; } = 0;
         public int GoldPitLevel { get; set; } = 15;
         public int AcademyLevel { get; set; } = 20;
         public int HydraHeads { get; set; } = 0;
