+++ SFSimulator.Core/Characters/ICharacterHelper.cs
@@ -6,15 +6,15 @@
         int GetDailyMissionReward(int characterLevel, bool experienceEvent, int hydraHeads);
         int GetExperienceForNextLevel(int characterLevel);
         QuestValue GetMinimumQuestValue(int characterLevel, ExperienceBonus experienceBonus, GoldBonus goldBonus);
-        float GetGoldFromGuardDuty(int level, GoldBonus? goldBonus, bool goldEvent);
-        float GetHourlyGoldPitProduction(int characterLevel, int goldPitLevel, bool goldEvent);
+        decimal GetGoldFromGuardDuty(int level, GoldBonus? goldBonus, bool goldEvent);
+        decimal GetHourlyGoldPitProduction(int characterLevel, int goldPitLevel, bool goldEvent);
         int GetExperienceRewardFromCalendar(int characterLevel, int rewardSize);
-        float GetGoldRewardFromCalendar(int characterLevel, int rewardSize);
+        decimal GetGoldRewardFromCalendar(int characterLevel, int rewardSize);
         int GetExperienceRewardFromArena(int characterLevel, bool experienceEvent);
-        float GetDailyGoldFromDiceGame(int characterLevel, IEnumerable<EventType> events);
+        decimal GetDailyGoldFromDiceGame(int characterLevel, IEnumerable<EventType> events);
         int GetDailyExperienceFromWheel(int characterLevel, IEnumerable<EventType> events, SpinAmountType spinAmount);
-        float GetDailyGoldFromWheel(int characterLevel, IEnumerable<EventType> events, SpinAmountType spinAmount);
+        decimal GetDailyGoldFromWheel(int characterLevel, IEnumerable<EventType> events, SpinAmountType spinAmount);
         int GetXPFromGuildFight(int characterLevel, IEnumerable<EventType> events);
-        float GetDailyGoldFromGemMine(int characterLevel, int gemMineLevel, int workers = 15);
+        decimal GetDailyGoldFromGemMine(int characterLevel, int gemMineLevel, int workers = 15);
     }
 }
