+++ SFSimulator.Core/Items/Item.cs
@@ -5,7 +5,7 @@
         public ItemType ItemType { get; set; }
 
         public ItemSourceType ItemSourceType { get; set; }
-        public float GoldValue { get; set; }
+        public decimal GoldValue { get; set; }
 
     }
 }
