+++ SFSimulator.Core/Quests/QuestChoosers/QuestChooser.cs
@@ -2,15 +2,15 @@
 {
     public class QuestChooser : IQuestChooser
     {
-        public double Multiplier { get; set; } = 0.85d;
+        public decimal Multiplier { get; set; } = 0.85M;
         public QuestChooser()
         {
         }
-        public QuestChooser(double multiplier)
+        public QuestChooser(decimal multiplier)
         {
             Multiplier = multiplier;
         }
-        public Quest ChooseBestQuest(IEnumerable<Quest> quests, QuestPriorityType questPriority, QuestChooserAI questChooserAI, float? hybridRatio = 0f)
+        public Quest ChooseBestQuest(IEnumerable<Quest> quests, QuestPriorityType questPriority, QuestChooserAI questChooserAI, decimal? hybridRatio = 0)
         {
             return questChooserAI switch
             {
@@ -19,7 +19,7 @@
                 _ => throw new ArgumentException($"Wrong {nameof(questChooserAI)} value"),
             };
         }
-        private Quest SimpleChooser(IEnumerable<Quest> quests, QuestPriorityType questPriority, float? hybridRatio)
+        private Quest SimpleChooser(IEnumerable<Quest> quests, QuestPriorityType questPriority, decimal? hybridRatio)
         {
             return questPriority switch
             {
@@ -39,14 +39,14 @@
             };
         }
 
-        private Quest SmartChooser(IEnumerable<Quest> quests, QuestPriorityType questPriority, float? hybridRatio)
+        private Quest SmartChooser(IEnumerable<Quest> quests, QuestPriorityType questPriority, decimal? hybridRatio)
         {
 
             if (questPriority == QuestPriorityType.Gold)
             {
                 var transformedQuests = quests.Select(q =>
                 {
-                    var multiplier = (10 - q.Time) / 2.5 * Multiplier;
+                    var multiplier = (10 - q.Time) / 2.5M * Multiplier;
                     var estimatedGold = q.Gold + q.MinimumValues.Gold * 5 * multiplier;
 
                     if (q.Item != null && q.Item.ItemSourceType == ItemSourceType.BeforeQuest)
@@ -61,7 +61,7 @@
             {
                 var transformedQuests = quests.Select(q =>
                 {
-                    var multiplier = (10 - q.Time) / 2.5 * Multiplier;
+                    var multiplier = (10 - q.Time) / 2.5M * Multiplier;
                     var estimatedExperience = q.Experience + q.MinimumValues.Experience * 5 * multiplier;
