+++ SFSimulator.Core/Quests/QuestFactory/QuestFactory.cs
@@ -11,15 +11,15 @@
             _itemGeneartor = itemGenerator;
             _random = random;
         }
-        public Quest Create(QuestValue minimumQuestValue, int characterLevel, double thirst, bool hasGoldScroll = false, float goldRune = 0, IEnumerable<EventType>? events = null, MountType mountType = MountType.Griffin)
+        public Quest Create(QuestValue minimumQuestValue, int characterLevel, decimal thirst, bool hasGoldScroll = false, decimal goldRune = 0, IEnumerable<EventType>? events = null, MountType mountType = MountType.Griffin)
         {
             var time = _random.Next(1, 5);
             return Create(minimumQuestValue, characterLevel, thirst, time, hasGoldScroll, goldRune, events, mountType);
         }
-        public Quest Create(QuestValue minimumQuestValue, int characterLevel, double thirst, int time, bool hasGoldScroll = false, float goldRune = 0, IEnumerable<EventType>? events = null, MountType mountType = MountType.Griffin)
+        public Quest Create(QuestValue minimumQuestValue, int characterLevel, decimal thirst, int time, bool hasGoldScroll = false, decimal goldRune = 0, IEnumerable<EventType>? events = null, MountType mountType = MountType.Griffin)
         {
             var convertedTime = _questHelper.GetTime(time, mountType);
-            var randomMultiplier = _random.NextDouble() + _random.Next(1, 5);
+            var randomMultiplier = (decimal)(_random.NextDouble() + _random.Next(1, 5));
             var gold = randomMultiplier * minimumQuestValue.Gold * time;
 
             var experience = (6 - randomMultiplier) * minimumQuestValue.Experience * time;
@@ -34,7 +34,7 @@
             experience = experience < 0 ? 0 : experience;
             gold = gold > 10000000 ? 10000000 : gold;
 
-            gold = hasGoldScroll ? gold * 1.1 * (1 + goldRune) : gold * (1 + goldRune);
+            gold = hasGoldScroll ? gold * 1.1M * (1 + goldRune) : gold * (1 + goldRune);
 
 
             //Truncate too long quests time and also their xp and gold
@@ -42,7 +42,7 @@
             {
                 gold *= thirst / convertedTime;
                 experience *= thirst / convertedTime;
-                convertedTime = (float)thirst;
+                convertedTime = thirst;
             }
 
             Item? item = GenerateItem(characterLevel);
@@ -50,11 +50,11 @@
             return new Quest(minimumQuestValue, gold, experience, convertedTime, item);
         }
 
-        public Quest CreateBonusQuest(QuestValue minimumQuestValue, int characterLevel, double thirst, bool hasGoldScroll = false, float goldRune = 0, IEnumerable<EventType>? events = null, MountType mountType = MountType.Griffin)
+        public Quest CreateBonusQuest(QuestValue minimumQuestValue, int characterLevel, decimal thirst, bool hasGoldScroll = false, decimal goldRune = 0, IEnumerable<EventType>? events = null, MountType mountType = MountType.Griffin)
         {
             var time = _random.Next(1, 5);
             var convertedTime = _questHelper.GetTime(time, mountType);
-            var randomMultiplier = _random.NextDouble() + _random.Next(1, 5);
+            var randomMultiplier = (decimal)(_random.NextDouble() + _random.Next(1, 5));
             var gold = randomMultiplier * minimumQuestValue.Gold * time;
 
             var experience = (6 - randomMultiplier) * minimumQuestValue.Experience * time;
@@ -70,12 +70,12 @@
             gold = gold < 0 ? 0 : gold;
             gold = gold > 10000000 ? 10000000 : gold;
 
-            gold = hasGoldScroll ? gold * 1.1 * (1 + goldRune) : gold * (1 + goldRune);
+            gold = hasGoldScroll ? gold * 1.1M * (1 + goldRune) : gold * (1 + goldRune);
 
             //Truncate too long quests time only (xp and gold for bonus quest is untouched)
             if (convertedTime > thirst)
             {
-                convertedTime = (float)thirst;
+                convertedTime = thirst;
             }
 
             Item? item = GenerateItem(characterLevel);
@@ -83,11 +83,11 @@
             return new Quest(minimumQuestValue, gold, experience, convertedTime, item);
         }
 
-        public Quest CreateTimeMachineQuest(QuestValue minimumQuestValue, double thirst, MountType mountType = MountType.Griffin)
+        public Quest CreateTimeMachineQuest(QuestValue minimumQuestValue, decimal thirst, MountType mountType = MountType.Griffin)
         {
             var time = _random.Next(1, 5);
             var convertedTime = _questHelper.GetTime(time, mountType);
-            var randomMultiplier = _random.NextDouble() + _random.Next(1, 5);
+            var randomMultiplier = (decimal)(_random.NextDouble() + _random.Next(1, 5));
             var gold = randomMultiplier * minimumQuestValue.Gold * time;
 
             var experience = (6 - randomMultiplier) * minimumQuestValue.Experience * time;
@@ -101,7 +101,7 @@
             {
                 gold *= thirst / convertedTime;
                 experience *= thirst / convertedTime;
-                convertedTime = (float)thirst;
+                convertedTime = thirst;
             }
 
             return new Quest(minimumQuestValue, gold, experience, convertedTime);
